re-organize attachment functions and updated comments doc
